refactor: #18-AuthService -> MemberService로 통합
