+++ src/main/java/freshtrash/freshtrashbackend/service/MemberService.java
@@ -6,7 +6,6 @@ import freshtrash.freshtrashbackend.exception.constants.ErrorCode;
 import freshtrash.freshtrashbackend.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -15,5 +14,32 @@ public class MemberService {
 
     public Member getMemberEntityByEmail(String email) {
         return memberRepository.findByEmail(email).orElseThrow(() -> new MemberException(ErrorCode.NOT_FOUND_MEMBER));
+    }
+
+    /**
+     * 회원 가입
+     */
+    public void registerMember(Member entity) {
+        if (checkEmailDuplication(entity.getEmail())) {
+            throw new MemberException(ErrorCode.ALREADY_EXISTS_EMAIL);
+        }
+        if (checkNicknameDuplication(entity.getNickname())) {
+            throw new MemberException(ErrorCode.ALREADY_EXISTS_NICKNAME);
+        }
+        memberRepository.save(entity);
+    }
+
+    /**
+     * 이메일 중복 체크
+     */
+    public boolean checkEmailDuplication(String email) {
+        return memberRepository.existsByEmail(email);
+    }
+
+    /**
+     * 닉네임 중복확인
+     */
+    public boolean checkNicknameDuplication(String nickname) {
+        return memberRepository.existsByNickname(nickname);
     }
 }
